+++ eduVPN/AppDelegate.swift
@@ -35,11 +35,11 @@ class AppDelegate: NSObject, NSApplicationDelegate {
       
         ValueTransformer.setValueTransformer(DurationTransformer(), forName: NSValueTransformerName(rawValue: "DurationTransformer"))
         
-        mainWindowController = NSStoryboard(name: NSStoryboard.Name(rawValue: "Main"), bundle: nil).instantiateController(withIdentifier: NSStoryboard.SceneIdentifier(rawValue: "MainWindowController")) as! MainWindowController
+        mainWindowController = NSStoryboard(name: "Main", bundle: nil).instantiateController(withIdentifier: "MainWindowController") as? MainWindowController
         mainWindowController.window?.makeKeyAndOrderFront(nil)
         
         // Adjust app name in menu and window
-        let appName = ServiceContainer.appName
+        let appName = ServiceContainer.appConfig.appName
         if appName != "eduVPN" {
             let fix: (NSMenuItem) -> Void = { menuItem in
                 menuItem.title = menuItem.title.replacingOccurrences(of: "eduVPN", with: appName)
@@ -67,14 +67,17 @@ class AppDelegate: NSObject, NSApplicationDelegate {
                     case .success:
                         NSApp.reply(toApplicationShouldTerminate: true)
                     case .failure(let error):
-                        NSApp.reply(toApplicationShouldTerminate: false)
-                        let alert = NSAlert(error: error)
-                        if let window = self.mainWindowController.window {
-                            alert.beginSheetModal(for: window) { (_) in
-                                
+                        if let alert = NSAlert(customizedError: error) {
+                            NSApp.reply(toApplicationShouldTerminate: false)
+                            if let window = self.mainWindowController.window {
+                                alert.beginSheetModal(for: window) { (_) in
+                                    
+                                }
+                            } else {
+                                alert.runModal()
                             }
                         } else {
-                            alert.runModal()
+                            NSApp.reply(toApplicationShouldTerminate: true)
                         }
                     }
                 }
@@ -104,7 +107,7 @@ class AppDelegate: NSObject, NSApplicationDelegate {
         window.makeKeyAndOrderFront(nil)
     }
     
-    private lazy var preferencesWindowController = NSStoryboard(name: NSStoryboard.Name(rawValue: "Main"), bundle: nil).instantiateController(withIdentifier: NSStoryboard.SceneIdentifier(rawValue: "PreferencesController")) as! NSWindowController
+    private lazy var preferencesWindowController = NSStoryboard(name: "Main", bundle: nil).instantiateController(withIdentifier: "PreferencesController") as! NSWindowController
     
     @objc @IBAction func showPreferences(_ sender: Any) {
         guard let window = mainWindowController.window, let preferencesWindow = preferencesWindowController.window else {
@@ -133,14 +136,15 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     
     private func updateStatusItemImage() {
         switch ServiceContainer.connectionService.state {
-        case .connecting:
-            statusItem?.image = #imageLiteral(resourceName: "disconnected-1")
+        case .connecting, .disconnecting:
+            statusItem?.image = #imageLiteral(resourceName: "Status-connecting")
+            NSApp.applicationIconImage = #imageLiteral(resourceName: "Icon-connecting")
         case .connected:
-            statusItem?.image = #imageLiteral(resourceName: "connected_bw")
-        case .disconnecting:
-            statusItem?.image = #imageLiteral(resourceName: "disconnected-1")
+            statusItem?.image = #imageLiteral(resourceName: "Status-connected")
+            NSApp.applicationIconImage = #imageLiteral(resourceName: "Icon-connected")
         case .disconnected:
-            statusItem?.image = #imageLiteral(resourceName: "disconnected-1")
+            statusItem?.image = #imageLiteral(resourceName: "Status-disconnected")
+            NSApp.applicationIconImage = #imageLiteral(resourceName: "Icon-disconnected")
         }
     }