+++ eduVPN/EnterProviderURLViewController.swift
@@ -62,8 +62,8 @@ class EnterProviderURLViewController: NSViewController {
         doneButton.isEnabled = false
         
         guard let url = validURL(), let _ = url.host else {
-            let alert = NSAlert(error: Error.invalidURL)
-            alert.beginSheetModal(for: self.view.window!) { (_) in
+            let alert = NSAlert(customizedError: Error.invalidURL)
+            alert?.beginSheetModal(for: self.view.window!) { (_) in
                 self.textField.isEnabled = true
             }
             return
@@ -78,7 +78,7 @@ class EnterProviderURLViewController: NSViewController {
 
 extension EnterProviderURLViewController: NSTextFieldDelegate {
     
-    override func controlTextDidChange(_ obj: Notification) {
+    func controlTextDidChange(_ obj: Notification) {
         doneButton.isEnabled = validURL() != nil
     }