+++ ChangeLog.md
@@ -4,9 +4,12 @@
 
 ### Version 1.1 (build 1)
 
+- new app icon which reflects connection state
+- new status bar icons
 - keychain support
 - update openvpn to 2.4.6
 - support for local config files
+- adding config files and URLs using drag-n-drop
 - improved handling of DNS settings
 - improved support for IPv6
 - certificate validity check before use